Split formattable tests into configuration specific files
